доработан код
