+++ index.html
@@ -63,7 +63,9 @@
           rows="4"
         ></textarea>
         <div class="add-form-row">
-          <button type="button" class="add-form-button" id="add-comment">Написать</button>
+          <button type="button" class="add-form-button" id="add-comment">
+            Написать
+          </button>
         </div>
       </div>
     </div>
