+++ modules/comments.js
@@ -1,14 +1,14 @@
 export let comments = [
   {
     name: "Глеб Фокин",
-    date: new Date(),
+    date: "12.02.22 12:18:",
     text: "Это будет первый комментарий на этой странице",
     likes: 3,
     isLiked: false,
   },
   {
     name: "Варвара Н.",
-    date: new Date(),
+    date: "13.02.22 19:22",
     text: "Мне нравится как оформлена эта страница! ❤",
     likes: 75,
     isLiked: true,
